requirements.txt
# Core
python-dotenv>=1.0.0
certifi>=2024.2.2
matplotlib
 
# ── Provider SDKs — install only what you need ──────────────────────────────
# Groq  (API_PROVIDER=groq)
groq>=0.9.0
 
# Gemini  (API_PROVIDER=gemini)
# google-generativeai>=0.5.0
 
# OpenAI / compatible endpoints  (API_PROVIDER=openai)
# openai>=1.30.0
 